mesh: update plot
