+++ fealpy/mesh/mesh_base/__init__.py
@@ -3,3 +3,4 @@ Provide ABCs for mesh
 """
 from .mesh import Mesh
 from .topology import Mesh1d, Mesh2d, Mesh3d
+from .plot import Plotable
